:building_construction: Migrated render functions
Migrated the renderChart functions to modules made for these functions
