+++ public/global.css
@@ -260,3 +260,11 @@ button:focus {
     top: 100%;
     left: 0;
 }
+
+/* rect{
+    opacity: .5;
+} */
+
+rect:hover{
+    opacity: 1;
+}
